requirements.txt
# Sistema Ótica - Dependências de Produção
# Gerado automaticamente - 2026-02-07

# Framework Web (versões fixas para estabilidade)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Segurança
pyjwt==2.8.0
python-dotenv==1.0.0

# Servidor ASGI
gunicorn==21.2.0
